+++ CHANGELOG.md
@@ -7,9 +7,12 @@ All notable changes to this project will be documented in this file.
 ### Added
 - `<impl Serializable>::length()`
 - `Serializer::with_capacity()`
+- `Serializer::write<impl Serializable>()`
+- `Deserializer::read<impl Serializable>()`
 ### Changed
 - `<impl Serializable>::try_to_bytes()` now allocates
   `<impl Serializable>::length()` bytes to the `Serializer` on creation
+- `kdf!` uses `$crate` local objects => remove `sha3` dependency in caller
 ### Fixed
 ### Removed
 ---